Handle invalid "before" queries
This does not change the behaviour of an "after" query errors.

These may occur when e.g. a `--targets` flag names a directory which was
added between `before` and `after`.

In this case, we treat all detected tests which match the `query` at
`after` as affected. This may lead to some over-triggering,
e.g. suppose we had the target:
* `//foo`

And the query:
`//foo/... + //bar/...`
If we created `//bar`, this would cause us to trigger `//foo` even
though it was not affected.

There's no way for us to programatically detect whether the query failed
because of a missing directory or some other reason, so we either need
to error to the user and leave them to manually resolve the issue, or
fall back to over-triggering targets. Over-triggering seems generally
safe (as there are for sure other cases where we'll over-trigger, e.g.
where source code had comment-only changes which led to the same
tests/binaries being produced), so we'll do the slightly more friendly
thing.