Allow shutting down as an analysis cache clear strategy
We've seen instances in the wild where shutdown can be minutes faster,
due to not invalidating a bunch of python whl_library repositories.

None of this is as ideal as cquery reporting just the information it got
from its particular run, but both of these heuristics may be useful.
Ideally we could get rid of them both at some point.