Allow  testing whether analysis cache clearing is necessary
From discussions at BazelCon, this _may_ be unnecessary, and it could be
that `bazel config` is the only command which leaks across invocations
now (which we can tolerate).

This adds a mode to run the cquery before and after the analysis cache
clear and diff them, validating whether the cache clear had any useful
effect.