Add option to skip analysis cache clearing
Discussions with the Bazel team at BazelCon suggested this was no longer
necessary, and having run extensively with
--compare-queries-around-analysis-cache-clear, we haven't seen any
issues with this.

I'm leaving the old options here as fallback in case we run into issues
in the future, so it's easy to work around them.

This should significantly speed up target determinator, as the second
set of queries should generally be much faster than when the whole
universe needed re-evaluating.