Use remote-jdk when running tests
This avoids dependency on local-jdk which may not always be present.

The test - changedBazelMajorVersion_native - still fails without
local-jdk because it uses bazel v3.7.1 which presumably does not support
flag --java_runtime_version. It might be best to add the remotejdk
setting in target-determinator-testdata repository to make the test
runner independent of bazel version inconsistencies.