Visibility test is more accurate
This test was previously falely passing because bazel-diff(er) was failing
because of a missing git commit in the testdata repo, not a behaviour of
the tool.

Query doesn't report visibility errors, but cquery does.

Make the testes instead expect their correct results, but the
target-determinator version test that a useful error is given instead.