Flags: discourage the use of absolute in-repo paths for bazel options
Document that the user should use the relative form for paths passed to bazel options via `--bazel-opts` and `--bazel-startup-opts`.

Target-determinator may query target information in git worktrees, which live in different directories. When this is the case, passing an absolute path in a bazel option, like `--bazel-startup-opts=--bazelrc=/my/repo.git/mybazelrc`, results in unexpected behaviors if the `mybazelrc` changed between the "before" and "after" commit.