fix(hash_cache): only take user execute bits into account
There are some cases where files are initially created with some
permissions (e.g. 775) but are then created by git with different
permissions (e.g. 755) when cloning, checking out a branch or creating a
worktree. Such differences don't matter to git nor to Bazel;
consequently, they should not change the hash of a source file in
target-determinator.

The execution bit for ther user (e.g. 744 vs 644) does matter, though,
and should be part of the hash.