Construct testdata at build time rather than cloning a repo
It turns out that making minor changes in the test data is prohibitively
complicated since the tests very much rely on the git shasums, and any
change will need to be reflected here. We also can't rebase the changes
since that would break historical builds.

Rather than rely on an external repo, instead, we now create the data
within our own build. The individual states we want to use are modelled
as separate directories within the integration test's `testdata` dir,
and each state is given a meaningful name.