Remove bazel-diff and bazel-differ comparison tests
These were useful while we were bootstrapping the project, but we are
not updating these any more, and we are unlikely to change the
behaviour of Target Determinator if an update causes some of these
comparison tests to fail.

For both of these reasons, it makes sense to remove the tests.