Canonicalize targets (#108)
This canonicalize all the labels in all attributes found that are label related using the output of `bazel mod dump_repo_mapping ""` for each commit.

This addresses #105

It doesn't have any integration tests yet because it would be much better to leverage https://github.com/bazel-contrib/target-determinator/pull/104 which is still in the works but it was tested against our internal monorepo and the reproducer in #105 

Label like attributes (string defined but has nodep = True) are special and handled as an exception as if they were labels, and thus also converted. (See [this thread](https://bazelbuild.slack.com/archives/CDCMRLS23/p1742821059464199))